+++ Base/BaseObject.cs
@@ -123,6 +123,9 @@ namespace WooCommerceNET.Base
 
         [DataMember(EmitDefaultValue = false)]
         public List<int> delete { get; set; }
+
+        [IgnoreDataMember]
+        public List<T> DeletedItems { get; set; }
     }
 
     public class WCItem<T>
@@ -139,37 +142,38 @@ namespace WooCommerceNET.Base
                 APIEndpoint = typeof(T).BaseType.GetRuntimeProperty("Endpoint").GetValue(null).ToString();
         }
 
-        public async Task<T> Get(int id, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Get(int id, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.GetRestful(APIEndpoint + "/" + id.ToString(), parms).ConfigureAwait(false));
         }
 
-        public async Task<T> Get(string email, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Get(string email, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.GetRestful(APIEndpoint + "/" + email, parms).ConfigureAwait(false));
         }
 
-        public async Task<List<T>> GetAll(Dictionary<string, string> parms = null)
+        public virtual async Task<List<T>> GetAll(Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<List<T>>(await API.GetRestful(APIEndpoint, parms).ConfigureAwait(false));
         }
 
-        public async Task<T> Add(T item, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Add(T item, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.PostRestful(APIEndpoint, item, parms).ConfigureAwait(false));
         }
 
+        [Obsolete("AddRange method is obsolete, please use UpdateRange for batch Add, Update, Delete.")]
         public async Task<BatchObject<T>> AddRange(BatchObject<T> items, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<BatchObject<T>>(await API.PostRestful(APIEndpoint + "/batch", items, parms).ConfigureAwait(false));
         }
 
-        public async Task<T> Update(int id, T item, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Update(int id, T item, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.PostRestful(APIEndpoint + "/" + id.ToString(), item, parms).ConfigureAwait(false));
         }
 
-        public async Task<T> UpdateWithNull(int id, object item, Dictionary<string, string> parms = null)
+        public virtual async Task<T> UpdateWithNull(int id, object item, Dictionary<string, string> parms = null)
         {
             if (API.GetType().Name == "RestAPI")
             {
@@ -194,12 +198,42 @@ namespace WooCommerceNET.Base
                 return API.DeserializeJSon<T>(await API.PostRestful(APIEndpoint + "/" + id.ToString(), item, parms).ConfigureAwait(false));
         }
 
-        public async Task<BatchObject<T>> UpdateRange(BatchObject<T> items, Dictionary<string, string> parms = null)
+        public virtual async Task<BatchObject<T>> UpdateRange(BatchObject<T> items, Dictionary<string, string> parms = null)
         {
-            return API.DeserializeJSon<BatchObject<T>>(await API.PostRestful(APIEndpoint + "/batch", items, parms).ConfigureAwait(false));
+            string json = await UpdateRangeRaw(items, parms);
+
+            if (items.delete == null || items.delete.Count == 0)
+                return API.DeserializeJSon<BatchObject<T>>(json);
+            else
+            {
+                BatchObject<T> batchResult = new BatchObject<T>();
+
+                if ((items.create == null || items.create.Count == 0) && (items.update == null || items.update.Count == 0))
+                {
+                    batchResult.DeletedItems = API.DeserializeJSon<List<T>>(json.Substring(json.IndexOf("[")).TrimEnd('}'));
+                }
+                else
+                {
+                    var pos = json.LastIndexOf("\"delete\":[");
+                    if (pos != -1)
+                    {
+                        batchResult = API.DeserializeJSon<BatchObject<T>>(json.Substring(0, pos - 1) + "}");
+                        batchResult.DeletedItems = API.DeserializeJSon<List<T>>(json.Substring(pos + 9).TrimEnd('}'));
+                    }
+                    else
+                        batchResult = API.DeserializeJSon<BatchObject<T>>(json);
+                }
+                
+                return batchResult;
+            }
         }
 
-        public async Task<T> Delete(int id, bool force = false, Dictionary<string, string> parms = null)
+        public virtual async Task<string> UpdateRangeRaw(BatchObject<T> items, Dictionary<string, string> parms = null)
+        {
+            return await API.PostRestful(APIEndpoint + "/batch", items, parms).ConfigureAwait(false);
+        }
+
+        public virtual async Task<T> Delete(int id, bool force = false, Dictionary<string, string> parms = null)
         {
             if (force)
             {
@@ -213,6 +247,7 @@ namespace WooCommerceNET.Base
             return API.DeserializeJSon<T>(await API.DeleteRestful(APIEndpoint + "/" + id.ToString(), parms).ConfigureAwait(false));
         }
 
+        [Obsolete("DeleteRange method is obsolete, please use UpdateRange for batch Add, Update, Delete.")]
         public async Task<string> DeleteRange(BatchObject<T> items, Dictionary<string, string> parms = null)
         {
             return await API.PostRestful(APIEndpoint + "/batch", items, parms).ConfigureAwait(false);
@@ -236,27 +271,27 @@ namespace WooCommerceNET.Base
             APIParentEndpoint = parentEndpoint;
         }
 
-        public async Task<T> Get(int id, int parentId, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Get(int id, int parentId, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.GetRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint + "/" + id.ToString(), parms).ConfigureAwait(false));
         }
 
-        public async Task<List<T>> GetAll(object parentId, Dictionary<string, string> parms = null)
+        public virtual async Task<List<T>> GetAll(object parentId, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<List<T>>(await API.GetRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint, parms).ConfigureAwait(false));
         }
 
-        public async Task<T> Add(T item, int parentId, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Add(T item, int parentId, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.PostRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint, item, parms).ConfigureAwait(false));
         }
 
-        public async Task<T> Update(int id, T item, int parentId, Dictionary<string, string> parms = null)
+        public virtual async Task<T> Update(int id, T item, int parentId, Dictionary<string, string> parms = null)
         {
             return API.DeserializeJSon<T>(await API.PostRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint + "/" + id.ToString(), item, parms).ConfigureAwait(false));
         }
 
-        public async Task<T> UpdateWithNull(int id, int parentId, object item, Dictionary<string, string> parms = null)
+        public virtual async Task<T> UpdateWithNull(int id, int parentId, object item, Dictionary<string, string> parms = null)
         {
             if (API.GetType().Name == "RestAPI")
             {
@@ -281,12 +316,42 @@ namespace WooCommerceNET.Base
                 return API.DeserializeJSon<T>(await API.PostRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint + "/" + id.ToString(), item, parms).ConfigureAwait(false));
         }
 
-        public async Task<BatchObject<T>> UpdateRange(int parentId, BatchObject<T> items, Dictionary<string, string> parms = null)
+        public virtual async Task<BatchObject<T>> UpdateRange(int parentId, BatchObject<T> items, Dictionary<string, string> parms = null)
         {
-            return API.DeserializeJSon<BatchObject<T>>(await API.PostRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint + "/batch", items, parms).ConfigureAwait(false));
+            string json = await UpdateRangeRaw( parentId, items, parms );
+
+            if ( items.delete == null || items.delete.Count == 0 )
+                return API.DeserializeJSon<BatchObject<T>>( json );
+            else
+            {
+                BatchObject<T> batchResult = new BatchObject<T>();
+
+                if ( ( items.create == null || items.create.Count == 0 ) && ( items.update == null || items.update.Count == 0 ) )
+                {
+                    batchResult.DeletedItems = API.DeserializeJSon<List<T>>( json.Substring( json.IndexOf( "[" ) ).TrimEnd( '}' ) );
+                }
+                else
+                {
+                    var pos = json.LastIndexOf( "\"delete\":[" );
+                    if ( pos != -1 )
+                    {
+                        batchResult = API.DeserializeJSon<BatchObject<T>>( json.Substring( 0, pos - 1 ) + "}" );
+                        batchResult.DeletedItems = API.DeserializeJSon<List<T>>( json.Substring( pos + 9 ).TrimEnd( '}' ) );
+                    }
+                    else
+                        batchResult = API.DeserializeJSon<BatchObject<T>>( json );
+                }
+
+                return batchResult;
+            }
         }
 
-        public async Task<string> Delete(int id, int parentId, bool force = false, Dictionary<string, string> parms = null)
+        public virtual async Task<string> UpdateRangeRaw(int parentId, BatchObject<T> items, Dictionary<string, string> parms = null)
+        {
+            return await API.PostRestful(APIParentEndpoint + "/" + parentId.ToString() + "/" + APIEndpoint + "/batch", items, parms).ConfigureAwait(false);
+        }
+
+        public virtual async Task<string> Delete(int id, int parentId, bool force = false, Dictionary<string, string> parms = null)
         {
             if (force)
             {