+++ WooCommerce/v1/Report.cs
@@ -119,7 +119,7 @@ namespace WooCommerceNET.WooCommerce.v1
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int product_id { get; set; }
+        public uint product_id { get; set; }
 
         /// <summary>
         /// Total number of purchases.