+++ WooCommerce/v1/Webhook.cs
@@ -15,7 +15,7 @@ namespace WooCommerceNET.WooCommerce.v1
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int id { get; set; }
+        public uint id { get; set; }
 
         /// <summary>
         /// A friendly name for the webhook. Defaults is Webhook created on lt;dategt;.
@@ -94,7 +94,7 @@ namespace WooCommerceNET.WooCommerce.v1
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// The delivery duration, in seconds.