+++ WooCommerce/v2/Order.cs
@@ -17,13 +17,13 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Parent order ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? parent_id { get; set; }
+        public uint? parent_id { get; set; }
 
         /// <summary>
         /// Order number. 
@@ -155,7 +155,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// User ID who owns the order. 0 for guests. Default is 0.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? customer_id { get; set; }
+        public uint? customer_id { get; set; }
 
         /// <summary>
         /// Customer’s IP address. 
@@ -437,7 +437,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Product name.
@@ -449,13 +449,13 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Product ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? product_id { get; set; }
+        public uint? product_id { get; set; }
 
         /// <summary>
         /// Variation ID, if applicable.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? variation_id { get; set; }
+        public uint? variation_id { get; set; }
 
         /// <summary>
         /// Quantity ordered.
@@ -595,7 +595,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// tax item id
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// tax item total
@@ -620,7 +620,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Shipping method name.
@@ -678,7 +678,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Fee name.
@@ -727,6 +727,12 @@ namespace WooCommerceNET.WooCommerce.v2
     }
 
     [DataContract]
+    public class OrderCouponLineMeta : WCObject.MetaData
+    {
+
+    }
+
+    [DataContract]
     public class OrderCouponLine
     {
         /// <summary>
@@ -734,7 +740,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Coupon code.
@@ -761,7 +767,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Meta data. See Order - Meta data properties
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public List<OrderMeta> meta_data { get; set; }
+        public List<OrderCouponLineMeta> meta_data { get; set; }
 
     }
 
@@ -773,7 +779,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Refund reason.