+++ WooCommerce/v2/OrderNote.cs
@@ -13,7 +13,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// The date the order note was created, in the site’s timezone.