+++ WooCommerce/v2/OrderRefund.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
+using WooCommerceNET.Base;
 
 namespace WooCommerceNET.WooCommerce.v2
 {
     [DataContract]
-    public class OrderRefund
+    public class OrderRefund : JsonObject
     {
         public static string Endpoint { get { return "refunds"; } }
 
@@ -14,7 +15,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// The date the order refund was created, in the site’s timezone. 
@@ -48,7 +49,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// User ID of user who created the refund.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int refunded_by { get; set; }
+        public uint? refunded_by { get; set; }
 
         /// <summary>
         /// Meta data. See Order refund - Meta data properties
@@ -85,7 +86,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Product name.
@@ -97,13 +98,13 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Product ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? product_id { get; set; }
+        public uint? product_id { get; set; }
 
         /// <summary>
         /// Variation ID, if applicable.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? variation_id { get; set; }
+        public uint? variation_id { get; set; }
 
         /// <summary>
         /// Quantity ordered.
@@ -115,7 +116,7 @@ namespace WooCommerceNET.WooCommerce.v2
         /// Tax class of product.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int tax_class { get; set; }
+        public int? tax_class { get; set; }
 
         /// <summary>
         /// Line subtotal (before discounts).