+++ WooCommerce/v3/Product.cs
@@ -17,7 +17,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// Product name.
@@ -378,7 +378,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// Product parent ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? parent_id { get; set; }
+        public uint? parent_id { get; set; }
 
         /// <summary>
         /// Optional note to send the customer after purchase.
@@ -442,6 +442,11 @@ namespace WooCommerceNET.WooCommerce.v3
         [DataMember(EmitDefaultValue = false)]
         public List<v2.ProductMeta> meta_data { get; set; }
 
+        /// <summary>
+        /// Container for error information, if any
+        /// </summary>
+        [DataMember(EmitDefaultValue = false)]
+        public ProductError error { get; set; }
     }
 
     [DataContract]
@@ -534,7 +539,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// The date the review was created, in the site’s timezone.
@@ -552,7 +557,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// Unique identifier for the product that the review belongs to.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? product_id { get; set; }
+        public uint? product_id { get; set; }
 
         /// <summary>
         /// Status of the review. Options: approved, hold, spam, unspam, transh and untrash. Defauls to approved.
@@ -591,4 +596,6 @@ namespace WooCommerceNET.WooCommerce.v3
         public bool? verified { get; set; }
     }
 
+    [DataContract]
+    public class ProductError : v2.ProductError { }
 }