+++ WooCommerce/v3/Variation.cs
@@ -15,7 +15,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// The date the variation was created, in the site’s timezone. 
@@ -262,6 +262,11 @@ namespace WooCommerceNET.WooCommerce.v3
         [DataMember(EmitDefaultValue = false)]
         public List<v2.VariationMeta> meta_data { get; set; }
 
+        /// <summary>
+        /// Container for error information, if any
+        /// </summary>
+        [DataMember(EmitDefaultValue = false)]
+        public VariationError error { get; set; }
     }
 
     [DataContract]
@@ -277,7 +282,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// Image ID.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public uint? id { get; set; }
 
         /// <summary>
         /// The date the image was created, in the site’s timezone. 
@@ -328,5 +333,8 @@ namespace WooCommerceNET.WooCommerce.v3
 
     [DataContract]
     public class VariationAttribute : v2.VariationAttribute { }
+
+    [DataContract]
+    public class VariationError : v2.VariationError { }
 
 }