+++ WordPress/v2/Posts.cs
@@ -49,7 +49,7 @@ namespace WooCommerce.NET.WordPress.v2
         /// Unique identifier for the object.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int id  { get; set; }
+        public uint id  { get; set; }
 
         /// <summary>
         /// URL to the object.
@@ -141,7 +141,7 @@ namespace WooCommerce.NET.WordPress.v2
         /// The ID for the author of the object.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int author { get; set; }
+        public uint author { get; set; }
 
         /// <summary>
         /// The excerpt for the object.